refactor(convert): move extract_block_style to convert/style/ submodule
